requirements.txt
streamlit==1.32.0
opencv-python-headless==4.9.0.80
numpy==1.26.4
pillow==10.2.0
scipy==1.12.0
